ELEC-268: Add DBC shim script (#33)
Add a shim script to generate a DBC from the ASCIIPB file using the old
CAN protocol. A new Makefile target (gen-dbc) is added. This script
handles all of the old CAN protocol (what I call "CANdlelight 1.0")
used for FSGP/ASC 2018 as far as I know, including:

    * Implicit Messages created as ACKs in the protocol layer
    * Treating the Battery V/T Message as a MUXed Message
    * Handling signed/unsigned typing of various Signals

The following messages (that are currently used) contain signed
signals, and are handled:

    * Drive Output:
        * throttle: int16_t
        * direction: int16_t
        * cruise_control: int16_t
        * mechanical_brake_state: int16_t
    * Cruise Target:
        * target speed: int16_t
    * Battery Aggregate V/C
        * voltage: uint16_t
        * current: int16_t
    * Motor Velocity:
        * vehicle_velocity_left: int16_t
        * vehicle_velocity_right: int16_t

In addition, this also updates the Travis CI configuration to deploy
the DBC (system_can.dbc) as part of a GitHub release.